refactor: Move `*Namespace` protocols to `_compliant`
They were added before the package existed and don't belong in `utils`
