+++ narwhals/_arrow/utils.py
@@ -13,7 +13,7 @@ from typing import overload
 import pyarrow as pa
 import pyarrow.compute as pc
 
-from narwhals.utils import _SeriesNamespace
+from narwhals._compliant.series import _SeriesNamespace
 from narwhals.utils import import_dtypes_module
 from narwhals.utils import isinstance_or_issubclass
